+++ src/LinkDotNet.StringBuilder/ValueStringBuilder.Append.cs
@@ -78,9 +78,27 @@ public ref partial struct ValueStringBuilder
     /// Appends a slice of memory.
     /// </summary>
     /// <param name="memory">The memory to add.</param>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public void Append(ReadOnlyMemory<char> memory)
     {
         Append(memory.Span);
+    }
+
+    /// <summary>
+    /// Appends a single character to the string builder.
+    /// </summary>
+    /// <param name="value">Character to add.</param>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public void Append(char value)
+    {
+        var newSize = bufferPosition + 1;
+        if (newSize > buffer.Length)
+        {
+            Grow(newSize);
+        }
+
+        buffer[bufferPosition] = value;
+        bufferPosition++;
     }
 
     /// <summary>